# src/Prowlarr.Api.V1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fetcharr",
    "description": "Fetcharr API docs",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Fetcharr/Fetcharr/blob/develop/LICENSE"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:9696"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v1/history/since": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/HistoryEventType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/indexer": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "indexerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/HistoryEventType"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HistoryEventType": {
        "enum": [
          "unknown",
          "releaseGrabbed",
          "indexerQuery",
          "indexerRss",
          "indexerAuth",
          "indexerInfo"
        ],
        "type": "string"
      },
      "HistoryResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "indexerId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "downloadId": {
            "type": "string",
            "nullable": true
          },
          "successful": {
            "type": "boolean"
          },
          "eventType": {
            "$ref": "#/components/schemas/HistoryEventType"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ]
}